move search to a search folder
