+++ client/src/components/article/article_show.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { DonutGraph } from './charts/donut_graph';
 import RadarChart  from './charts/radar';
-import WikiSearch from '../search';
+import WikiSearchContainer from '../search/search_container';
 
 export class ArticleShow extends React.Component {
     constructor(props) {
@@ -14,7 +14,7 @@ export class ArticleShow extends React.Component {
     render() {
         return (
             <div className="article-show-page-container">
-                <WikiSearch />
+                <WikiSearchContainer />
                 <div className="article-show-charts">
                     <RadarChart />
                     <DonutGraph articleUrl={this.state.articleUrl}/>
